+++ test/widget/chat_direct_link_chat_test.dart
@@ -45,7 +45,6 @@ import 'package:messenger/provider/hive/contact.dart';
 import 'package:messenger/provider/hive/draft.dart';
 import 'package:messenger/provider/hive/gallery_item.dart';
 import 'package:messenger/provider/hive/media_settings.dart';
-import 'package:messenger/provider/hive/preferences.dart';
 import 'package:messenger/provider/hive/session.dart';
 import 'package:messenger/provider/hive/user.dart';
 import 'package:messenger/routes.dart';
@@ -131,6 +130,9 @@ void main() async {
       Get.put(AuthService(AuthRepository(graphQlProvider), sessionProvider));
   await authService.init();
 
+  router = RouterState(authService);
+  router.provider = MockPlatformRouteInformationProvider();
+
   var galleryItemProvider = Get.put(GalleryItemHiveProvider());
   await galleryItemProvider.init();
   await galleryItemProvider.clear();
@@ -159,11 +161,6 @@ void main() async {
   await chatItemHiveProvider.clear();
   var credentialsProvider = ChatCallCredentialsHiveProvider();
   await credentialsProvider.init();
-  var preferences = Get.put(PreferencesHiveProvider());
-  await preferences.init();
-
-  router = RouterState(authService);
-  router.provider = MockPlatformRouteInformationProvider();
 
   Widget createWidgetForTesting({required Widget child}) {
     FlutterError.onError = ignoreOverflowErrors;