+++ test/widget/chat_reply_message_test.dart
@@ -46,7 +46,6 @@ import 'package:messenger/provider/hive/contact.dart';
 import 'package:messenger/provider/hive/draft.dart';
 import 'package:messenger/provider/hive/gallery_item.dart';
 import 'package:messenger/provider/hive/media_settings.dart';
-import 'package:messenger/provider/hive/preferences.dart';
 import 'package:messenger/provider/hive/session.dart';
 import 'package:messenger/provider/hive/user.dart';
 import 'package:messenger/routes.dart';
@@ -276,6 +275,9 @@ void main() async {
       AuthService(AuthRepository(graphQlProvider), SessionDataHiveProvider());
   await authService.init();
 
+  router = RouterState(authService);
+  router.provider = MockPlatformRouteInformationProvider();
+
   var galleryItemProvider = Get.put(GalleryItemHiveProvider());
   await galleryItemProvider.init();
   await galleryItemProvider.clear();
@@ -300,11 +302,6 @@ void main() async {
   await backgroundProvider.init();
   var credentialsProvider = ChatCallCredentialsHiveProvider();
   await credentialsProvider.init();
-  var preferences = Get.put(PreferencesHiveProvider());
-  await preferences.init();
-
-  router = RouterState(authService);
-  router.provider = MockPlatformRouteInformationProvider();
 
   var messagesProvider = Get.put(ChatItemHiveProvider(
     const ChatId('0d72d245-8425-467a-9ebd-082d4f47850b'),