+++ CHANGELOG.md
@@ -31,6 +31,8 @@ All user visible changes to this project will be documented in this file. This p
         - Searching. ([#260], [#259])
     - User page:
         - Blacklisting. ([#234], [#229])
+- Desktop:
+    - Persist window's size and position ([#274], [#263])
 
 ### Changed
 
@@ -58,8 +60,6 @@ All user visible changes to this project will be documented in this file. This p
         - Alphabetical and last seen sorting. ([#235], [#226])
     - Chat info page:
         - Redesigned chat info page. ([#265], [#256])    
-- Desktop:
-    - Persist window's size and position ([#274], [#263])
 
 ### Fixed